+++ CHANGELOG.md
@@ -80,6 +80,7 @@
 - Verify quiz session status without having a second learning break !1018 #673 @svenpascal
 - Show message and redirect page after deleting an account !1028 #533 @alex.simm
 - Fixed the createThread call inside the ForumRestController !1031 @ctwx_ok
+- Remove forum topic subscriptions when leaving group !1020 #593 @alex.simm
 
 ## Refactoring
 - Refactored profile from WorkGroupModel to WorkGroupGateway !898 #9 @svenpascal